rename primary network
