+++ crates/consensus/primary/src/primary.rs
@@ -7,7 +7,7 @@ use crate::{
     certificate_fetcher::CertificateFetcher,
     certifier::Certifier,
     consensus::{ConsensusRound, LeaderSchedule},
-    peer_handler::{PrimaryReceiverHandler, WorkerReceiverHandler},
+    network::{PrimaryReceiverHandler, WorkerReceiverHandler},
     proposer::Proposer,
     state_handler::StateHandler,
     synchronizer::Synchronizer,
